requirements.txt
aiohttp==3.11.2
websockets==14.1
PyJWT==2.9.0